+++ CHANGELOG.md
@@ -1,1 +1,2 @@
 - `firebase ext:dev:init` now creates a placeholder `CHANGELOG.md` file.
+- Fix Auth Emulator not updating lastLoginAt with signInWithPassword, causing some "token revoked" errors.